% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R, R/hecfdar-package.R
\docType{package}
\name{hecfdar-package}
\alias{hecfdar-package}
\title{hecfdar: R interface to the HEC-FDA C++17 compute core}
\description{
Internal cpp11 bindings are registered via useDynLib and reached
through \code{asNamespace("hecfdar")} in the package's test suite.

R interface to a C++17 port of the HEC-FDA flood damage analysis compute engine. Reproduces the upstream computations with seeded reproducibility shared with the Python package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cameronbracken/hecfda}
  \item \url{https://cameronbracken.github.io/hecfda/}
  \item Report bugs at \url{https://github.com/cameronbracken/hecfda/issues}
}

}
\author{
\strong{Maintainer}: Cameron Bracken \email{cameron.bracken@pm.me}

Authors:
\itemize{
  \item Cameron Bracken \email{cameron.bracken@pm.me}
}

}
\keyword{internal}
